infra: move Model.deploy unit tests to separate file
